Crate bee_network

Networking layer for the Bee framework.

Modules

config

conns

interaction

network

peers

protocols

transport

Structs

Keypair
An Ed25519 keypair.

Multiaddr
Representation of a Multiaddr.

NetworkConfig
The network configuration.

NetworkConfigBuilder
A network configuration builder.

NetworkController
A controller for the networking layer, that allows to issue various commands, e.g. sending a message to a peer.

PeerId
Identifier of a peer of the network.

PeerInfo
Additional information about a peer.

Enums

Command
Describes the commands accepted by the networking layer.

Event
Describes the events produced by the networking layer.

Origin
Describes direction of an established connection.

PeerRelation
Describes the relation with a peer.

Protocol
Protocol describes all possible multiaddress protocols.

PublicKey
The public key of a node's identity keypair.

Statics

MAX_UNKNOWN_PEERS

NETWORK_ID

RECONNECT_INTERVAL_SECS

Traits

ShortId
A trait specifically there to create shorter peer ids for better readability in logs and user interfaces.

Functions

init
Initializes the networking layer.

Type Definitions

MessageReceiver
A shorthand for an unbounded channel receiver.

MessageSender
A shorthand for an unbounded channel sender.

NetworkListener
A type that receives any event published by the networking layer.