Trait bee_network::ShortId

pub trait ShortId where
    Self: ToString, {
    const SHORT_LENGTH: usize;

    fn short(&self) -> String;
}

A trait specifically there to create shorter peer ids for better readability in logs and user interfaces.

Associated Constants

const SHORT_LENGTH: usize

The length of the shortened peer id.

Loading content...
Required methods

fn short(&self) -> String

Creates a shorter - more readable - id from the original.

Loading content...